Minor: sets up basic structure for files for sign up components
