+++ src/components/index.ts
@@ -2,3 +2,4 @@ export * from './header';
 export * from './profile';
 export * from './sidebar';
 export * from './sidebar';
+export * from './signin';
