+++ src/components/signin/index.ts
@@ -1,0 +1,3 @@
+import { SignInContainer } from './signin-container';
+
+export { SignInContainer };
