+++ src/components/signin/signin-container.tsx
@@ -1,0 +1,9 @@
+const SignInContainer = () => {
+  return (
+    <div>
+      <h1>Sign In</h1>
+    </div>
+  );
+};
+
+export { SignInContainer };
